requirements.txt
fastapi==0.75.2
python-dotenv==0.21.0
numpy==1.23.1
uvicorn==0.17.6
transformers[torch]